rag-backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sentence-transformers==5.1.0
faiss-cpu==1.12.0
transformers==4.56.0
torch==2.8.0
pydantic==2.11.7
python-multipart==0.0.20
openai==1.52.0
python-dotenv==1.0.1
